We're currently at stage where we can create any numbers of areas and travel around! Map updated to a newer version, added a new "interior" map, added a switching map function.
Since we want different levels, different places to handle missions or just to explore we need to handle events, objects, monsters and npc at every single map. Changed the mapTileNum to three-dimensional array. Changed the Obj, Npc, Monster and iTile to three-dimensional array. All the for-loops in our update methods are now using the second dimension length.

In AssetSetter we're not using one more variable which is mapNum. So if you're placing an object you can now choose on which map you want it.

In CollisionChecker we implemented the same technique. Since we need to check on other maps too.

In Entity, we check on which map is the item dropped.

In EventRect, we also check on which map is the event happening plus we added a new Hut Entry/Exit event. (Map Switching).

In KeyHandler, we can now update depending on what map we're currently playing.