Implementing Pathfinding Algorithm so NPCs & Monsters can search the path and reach a specified location/target.
Three possible scenarios for our game:
- Guiding NPC to a different location
- NPC following you
- Making monsters start attacking you/ chasing you

The used algorithm is "A* Search Algorithm". Widely used in computer science and game development. There is ton explained in internet, so I'll let you do the research.

Handled stuff:
In Entity is created a new house-keeping method "checkCollision" and moved the code from the "update" method. We created another method called "searchPath" handling the entity guidelines.

Adjusted MON_GreenSlime behaviour. There is now a few conditions. First one is that whenever a player attacks the monster, it immediately becomes aggro (follows and attacks the player). Second one is that whenever a player is nearby a monster (10 tiles distance set) a randomizer is set and if it hits 50% the monster gets aggro again. Third one is that when the player moves a certain amount of tile distance (14 set), monster aggro is off.

Adjusted NPC_OldMan behaviour. solidArea is changed. speed changed. Two different conditions. NPC can now either follow the player or can be set to go somewhere whenever you speak to him.

In TileManager is declared new boolean "drawPath". Whenever NPC or Monster algo is triggered we can see path displayed on our screen. Showing is where it's going.