Implementing Day/Night Cycle. Also added a OBJ_Tent so when it gets night you can sleep. When you sleep you restore health and mana.
Handled:
In AssetSetter placed an OBJ_Tent on the map.

In EM_Lighting created the day/night cycle, and it's time. Plus added a "debug" (Day/Dusk/Night/Dawn) text on the screen.

In GamePanel added a new gameState.

Added new OBJ_Tent, it's default values and what happens when used.

In Player created a new getSleepingImage method, so now when you go to sleep it shows a tent on the map, and when you wake up it goes back to normal.

In UI class handled the Sleep State.